--- src/components/EncounterTable/EncounterTable.module.scss
@@ -1,39 +1,0 @@
-@use "@/styles/constants";
-
-.encounter-table {
-    --sprite-size: 120px;
-
-    .header {
-        font-size: 1.7rem;
-    }
-
-    .table {
-        border-radius: 5px;
-
-        .table-header {
-            font-size: 1.4rem;
-            padding: 1rem;
-        }
-
-        .row {
-            .table-element {
-                background-color: constants.$accent-color;
-                font-size: 1.2rem;
-                padding: 0.5rem 1rem;
-                text-align: center;
-                white-space: pre-line;
-
-                .pokemon {
-                    align-items: center;
-                    display: flex;
-
-                    .sprite {
-                        height: var(--sprite-size);
-                        position: relative;
-                        width: var(--sprite-size);
-                    }
-                }
-            }
-        }
-    }
-}
