+++ src/components/EncountersAccordion/EncountersAccordion.module.scss
@@ -8,18 +8,27 @@
     --name-font-size: 1rem;
     --name-spacing: 4px;
     --sprite-size: 90px;
+    --table-border-size: 1px;
+    --table-el-x-padding: 1rem;
+    --table-el-y-padding: 0.5rem;
+    --table-font-size: 1rem;
     --tier-border-rad: 2px;
     --tier-font-size: 0.8rem;
+    --tier-x-padding: 0.2rem;
+    --tier-y-padding: 0.1rem;
     --tier-size: 1.2rem;
     --transition-time: 0.2s;
     --type-ratio: 8/3;
     --type-height: 12px;
+    --x-padding: 1rem;
 
     width: var(--accordion-width);
 
     .header {
         align-items: center;
+        background: constants.$accent-color;
         display: flex;
+        padding: 0 var(--x-padding);
         position: relative;
         width: 100%;
 
@@ -44,17 +53,12 @@
                 }
 
                 .tier {
-                    align-items: center;
                     border-radius: var(--tier-border-rad);
-                    display: flex;
                     font-family: constants.$tier-font;
                     font-size: var(--tier-font-size);
                     font-weight: bold;
-                    height: var(--tier-size);
-                    justify-content: center;
                     line-height: var(--tier-font-size);
-                    padding: var(--tier-padding);
-                    width: var(--tier-size);
+                    padding: var(--tier-y-padding) var(--tier-x-padding);
                 }
 
                 .type {
@@ -82,6 +86,24 @@
             transform: rotate(180deg);
         }
     }
+
+    .body {
+        border-top: var(--table-border-size) solid constants.$text-color;
+        padding: none;
+        width: 100%;
+
+        .table-row {
+            .table-header,
+            .table-element {
+                background-color: constants.$accent-color;
+                font-size: var(--table-font-size);
+                padding: var(--table-el-y-padding) var(--table-el-x-padding);
+                text-align: center;
+                white-space: pre-line;
+                width: calc(100% / 3);
+            }
+        }
+    }
 }
 
 @media screen and (min-width: constants.$tablet) {
@@ -91,6 +113,8 @@
         --name-spacing: 8px;
         --tier-font-size: 1rem;
         --tier-size: 1.4rem;
+        --tier-x-padding: 0.4rem;
+        --tier-y-padding: 0.2rem;
     }
 }
