+++ src/styles/globals.scss
@@ -4,7 +4,7 @@
 @import url("https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,400;0,700;1,400&family=Chivo+Mono:wght@700&display=swap");
 
 #app {
-    color: white;
+    color: constants.$text-color;
     overflow-x: hidden;
 }
